BD-103-work-day-entity-의-spent-hour-을-int-16-double-로-수정

+++ Rlog/CoreData/CoreDataManager.swift
@@ -115,7 +115,7 @@ extension CoreDataManager {
     }
 
     // MARK: - WORKDAY CRUD
-    func createWorkday(of workspace: WorkspaceEntity, weekDay: Int16, yearInt: Int16, monthInt: Int16, dayInt: Int16, startTime: String, endTime: String, spentHour: Int16, workDayType: Int16) {
+    func createWorkday(of workspace: WorkspaceEntity, weekDay: Int16, yearInt: Int16, monthInt: Int16, dayInt: Int16, startTime: String, endTime: String, spentHour: Double, workDayType: Int16) {
         let workday = WorkDayEntity(context: context)
         workday.workspace = workspace
         workday.id = UUID()
@@ -159,7 +159,7 @@ extension CoreDataManager {
         return result ?? []
     }
 
-    func editWorkday(of workday: WorkDayEntity, weekDay: Int16, yearInt: Int16, monthInt: Int16, dayInt: Int16, startTime: String, endTime: String, spentHour: Int16, hasDone: Bool) {
+    func editWorkday(of workday: WorkDayEntity, weekDay: Int16, yearInt: Int16, monthInt: Int16, dayInt: Int16, startTime: String, endTime: String, spentHour: Double, hasDone: Bool) {
         workday.weekDay = weekDay
         workday.workDayType = workDayType
         workday.yearInt = yearInt
