+++ Rlog/Model/CoreData/WorkDayEntity+CoreDataProperties.swift
@@ -25,7 +25,7 @@ extension WorkDayEntity {
     @NSManaged public var startMinute: Int16
     @NSManaged public var endHour: Int16
     @NSManaged public var endMinute: Int16
-    @NSManaged public var spentHour: Int16
+    @NSManaged public var spentHour: Double
 
     @NSManaged public var hasDone: Bool
     @NSManaged public var workDayType: Int16
